+++ database/migrations/2022_11_29_053038_create_delivery_notes_table.php
@@ -39,9 +39,9 @@ return new class () extends Migration {
 
             $table->boolean('delivery_locked')->default(false);
 
-            $table->unsignedSmallInteger('address_id')->index()->nullable();
+            $table->unsignedInteger('address_id')->index()->nullable();
             $table->foreign('address_id')->references('id')->on('addresses');
-            $table->unsignedSmallInteger('delivery_country_id')->index()->nullable();
+            $table->unsignedInteger('delivery_country_id')->index()->nullable();
             $table->foreign('delivery_country_id')->references('id')->on('countries');
 
             $table->decimal('weight', 16)->nullable()->default(0);
